The What is Cosplay? webpage is complete
Using the index page as a template i created the what is cosplay
webpage. The only differences made were to the main body. The
announcements were substituted for the definition of cosplay and the
allocated cosplay links, plus I also changed the 'check out these
cosplayers' icons down the bottom of the page.