Store page complete!
Using the same method as with the other web pages i completed this page
by using the index.html page as a template. In this page i managed to
use a table element to create the list of second hand costume prices
with smaller text below it in order to make this deterring information
less visible.