CSS code
I have completed the CSS and have applied it to all the CossiePlay web
pages. The result looks good, sleek and professional considering that I
was only allowed to use the images provided and what i had made myself
in photoshop. I know that my previous code is correct as the layout and
image sizes are consistent throughout the entirety of the Cossieplay
website.